+++ src/resources/views/order/show/_extra_content.blade.php
@@ -1,0 +1,3 @@
+{{-- Override this view in your applications by creating --}}
+{{-- `resources/views/vendor/vanilo/order/show/_extra_content.blade.php` --}}
+{{-- to include additional content, cards, blocks, info, etc regarding the order --}}