+++ src/Providers/EventServiceProvider.php
@@ -18,4 +18,8 @@ use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvi
 class EventServiceProvider extends ServiceProvider
 {
     protected $listen = [];
+
+    protected function configureEmailVerification(): void
+    {
+    }
 }