+++ Changelog.md
@@ -11,6 +11,7 @@
 - Added support for shipping line items
 - Added cards with proper info to the shipping method's show page
 - Added the triggering of the `OrderStateChangedToPending` event when changing the order status to `pending` in the Admin
+- Added the `order/show/_extra_content` partial so that apps can add further elements to the bottom of the order show page
 - Fixed an error with Carbon 3 on the coupon show page
 - Fixed a bug where orders with multiple payments would only show the last payment in the modal